feat(json-files): add header replace/delete criteria
Added specific vendor/client files as well as overall files for updateing and deleting header names. This is to ensure everything is standardized.

BREAKING CHANGE: Some headers may overlap with some clients, there will be cases where we have a header renamed where we may not have wanted it to be renamed. These are rare case by case basis and should be manually changed.