feat(clients/api): add householding and padding
Added householding for I360 data, this will likely be adjusted to run for all clients that require it (L2, DataTrust, etc) rather than running in I360 alone. Padding was added for certain fields for many client specific tasks.

BREAKING CHANGE: There are possibilities for householding to go over the desired limit of 4 duplicate cases. It is not confirmed that more than 4 cases will appear, but in the case that they do a stop has not been put in place to remove those records from the sample file.