feat(db/api): add db
Added a database with tables mapped to move the json format to a database that can be used to update information. This was done so that the end user can update the information in the database as needed through the ui.

BREAKING CHANGE: If you do not have an updated .env file and you attempt to run the program, it will fail. This is because i explicitely made it so that the db_uri value is required, meaning a database connection will be required to use this program going forward.