fix(ui,-api): fix bug for vendor and file swap, restructure save for readability, add ui functionality, rem sample frame
This update addresses a few key issues in the ui and api interaction. Buttons have been added to clear file paths and default to mixed projects for landline and cell. Previously these options could not be reverted once they were selected, making it so that the user would need to restart the program to clear those fields. The save file function has been changed to be much shorter and much more readable. All sample frame testing was removed. Client was swapped to Vendor as to identify appropriately what we are choosing by. This is because our clients may choose different vendors, and some clients are their own vendor.

BREAKING CHANGE: There has been a few scenarios where the join path does not set properly when swapping vendors. This issue was only able to be recreated once, but there was no identifiable cause. This causes the program to see the join file path, but not read the file. The consequence is that a stratification cannot be run due to missing file. This has been noticed to happen when swapping between Tarrance and Baselice.